EDU-720 Use main-track-editor in multitrack plugin
