+++ src/plugins/interactive-media/interactive-media-editor.js
@@ -1,11 +1,9 @@
 import by from 'thenby';
 import classNames from 'classnames';
-import Logger from '../../common/logger.js';
 import { useTranslation } from 'react-i18next';
 import validation from '../../ui/validation.js';
 import cloneDeep from '../../utils/clone-deep.js';
 import Timeline from '../../components/timeline.js';
-import { handleError } from '../../ui/error-helper.js';
 import { removeItemAt } from '../../utils/array-utils.js';
 import { useOnComponentMounted } from '../../ui/hooks.js';
 import MediaPlayer from '../../components/media-player.js';
@@ -14,24 +12,17 @@ import React, { Fragment, useEffect, useState } from 'react';
 import DeleteButton from '../../components/delete-button.js';
 import MarkdownInput from '../../components/markdown-input.js';
 import InteractiveMediaInfo from './interactive-media-info.js';
-import { getResourceType } from '../../utils/resource-utils.js';
-import ResourcePicker from '../../components/resource-picker.js';
 import { useService } from '../../components/container-context.js';
+import { Button, Divider, Form, Input, Spin, Tooltip } from 'antd';
 import { sectionEditorProps } from '../../ui/default-prop-types.js';
+import MainTrackEditor from '../../components/main-track-editor.js';
 import { useNumberFormat } from '../../components/locale-context.js';
 import ObjectWidthSlider from '../../components/object-width-slider.js';
-import MediaRangeSelector from '../../components/media-range-selector.js';
-import { Button, Divider, Form, Input, Radio, Spin, Switch, Tooltip } from 'antd';
 import { CheckOutlined, LeftOutlined, PlusOutlined, RightOutlined } from '@ant-design/icons';
-import { storageLocationPathToUrl, urlToStorageLocationPath } from '../../utils/storage-utils.js';
+import { MEDIA_SCREEN_MODE, MEDIA_SOURCE_TYPE, RESOURCE_TYPE } from '../../domain/constants.js';
 import { analyzeMediaUrl, determineMediaDuration, ensureValidMediaPosition, formatMediaPosition } from '../../utils/media-utils.js';
-import { CDN_URL_PREFIX, MEDIA_ASPECT_RATIO, MEDIA_SCREEN_MODE, MEDIA_SOURCE_TYPE, RESOURCE_TYPE } from '../../domain/constants.js';
-
-const logger = new Logger(import.meta.url);
 
 const FormItem = Form.Item;
-const RadioGroup = Radio.Group;
-const RadioButton = Radio.Button;
 
 const formItemLayout = {
   labelCol: { span: 4 },
@@ -56,7 +47,7 @@ function InteractiveMediaEditor({ content, onContentChanged }) {
   const [selectedChapterIndex, setSelectedChapterIndex] = useState(0);
   const [selectedChapterFraction, setSelectedChapterFraction] = useState(0);
   const [isDeterminingDuration, setIsDeterminingDuration] = useState(false);
-  const { sourceType, sourceUrl, playbackRange, chapters, copyrightNotice, width, aspectRatio, showVideo } = content;
+  const { sourceType, sourceUrl, playbackRange, chapters, width, showVideo } = content;
 
   const playbackDuration = (playbackRange[1] - playbackRange[0]) * sourceDuration;
 
@@ -124,62 +115,6 @@ function InteractiveMediaEditor({ content, onContentChanged }) {
     chapter.startPosition = newStartPosition;
     const newChapters = [...chapters];
     changeContent({ chapters: newChapters });
-  };
-
-  const handleSourceTypeChange = event => {
-    const { value } = event.target;
-    setSelectedChapterIndex(0);
-    changeContent({
-      sourceType: value,
-      sourceUrl: '',
-      range: [0, 1],
-      showVideo: false,
-      copyrightNotice: ''
-    });
-  };
-
-  const handleSourceUrlChangeComplete = async value => {
-    const { sanitizedUrl, duration, range, resourceType, error } = await getMediaInformation(value);
-    setSourceDuration(duration);
-
-    const newCopyrightNotice = sourceType === MEDIA_SOURCE_TYPE.youtube
-      ? t('common:youtubeCopyrightNotice', { link: value })
-      : copyrightNotice;
-
-    changeContent({
-      sourceUrl: sourceType !== MEDIA_SOURCE_TYPE.internal ? sanitizedUrl : value,
-      showVideo: resourceType === RESOURCE_TYPE.video || resourceType === RESOURCE_TYPE.unknown,
-      playbackRange: range,
-      copyrightNotice: newCopyrightNotice
-    });
-
-    if (error) {
-      handleError({ error, logger, t });
-    }
-  };
-
-  const handleSourceUrlChange = event => {
-    changeContent({ sourceUrl: event.target.value });
-  };
-
-  const handleSourceUrlBlur = event => {
-    handleSourceUrlChangeComplete(event.target.value);
-  };
-
-  const handlePlaybackRangeChange = newRange => {
-    changeContent({ playbackRange: newRange });
-  };
-
-  const handleAspectRatioChanged = event => {
-    changeContent({ aspectRatio: event.target.value });
-  };
-
-  const handleShowVideoChanged = newShowVideo => {
-    changeContent({ showVideo: newShowVideo });
-  };
-
-  const handleCopyrightNoticeChanged = event => {
-    changeContent({ copyrightNotice: event.target.value });
   };
 
   const handleWidthChanged = newValue => {
@@ -273,16 +208,8 @@ function InteractiveMediaEditor({ content, onContentChanged }) {
     changeContent({ chapters: newChapters });
   };
 
-  const renderPlaybackRangeInfo = () => {
-    const from = playbackRange[0] !== 0
-      ? formatMediaPosition({ formatPercentage, position: playbackRange[0], duration: sourceDuration })
-      : 'start';
-
-    const to = playbackRange[1] !== 1
-      ? formatMediaPosition({ formatPercentage, position: playbackRange[1], duration: sourceDuration })
-      : 'end';
-
-    return t('playbackRangeInfo', { from, to });
+  const handleMainTrackContentChange = changedContent => {
+    changeContent({ ...changedContent });
   };
 
   const renderAnswer = (answer, index) => (
@@ -312,75 +239,12 @@ function InteractiveMediaEditor({ content, onContentChanged }) {
   return (
     <div className="InteractiveMediaEditor">
       <Form layout="horizontal">
-        <FormItem label={t('common:source')} {...formItemLayout}>
-          <RadioGroup value={sourceType} onChange={handleSourceTypeChange}>
-            <RadioButton value={MEDIA_SOURCE_TYPE.external}>{t('common:externalLink')}</RadioButton>
-            <RadioButton value={MEDIA_SOURCE_TYPE.internal}>{t('common:internalCdn')}</RadioButton>
-            <RadioButton value={MEDIA_SOURCE_TYPE.youtube}>{t('common:youtube')}</RadioButton>
-          </RadioGroup>
-        </FormItem>
-        {sourceType === MEDIA_SOURCE_TYPE.external && (
-          <FormItem label={t('common:externalUrl')} {...formItemLayout} {...validation.validateUrl(sourceUrl, t)} hasFeedback>
-            <Input value={sourceUrl} onChange={handleSourceUrlChange} onBlur={handleSourceUrlBlur} />
-          </FormItem>
-        )}
-        {sourceType === MEDIA_SOURCE_TYPE.internal && (
-          <FormItem label={t('common:internalUrl')} {...formItemLayout}>
-            <div className="u-input-and-button">
-              <Input
-                addonBefore={CDN_URL_PREFIX}
-                value={sourceUrl}
-                onChange={handleSourceUrlChange}
-                onBlur={handleSourceUrlBlur}
-                />
-              <ResourcePicker
-                url={storageLocationPathToUrl(sourceUrl)}
-                onUrlChange={url => handleSourceUrlChangeComplete(urlToStorageLocationPath(url))}
-                />
-            </div>
-          </FormItem>
-        )}
-        {sourceType === MEDIA_SOURCE_TYPE.youtube && (
-          <FormItem label={t('common:youtubeUrl')} {...formItemLayout} {...validation.validateUrl(sourceUrl, t)} hasFeedback>
-            <Input value={sourceUrl} onChange={handleSourceUrlChange} onBlur={handleSourceUrlBlur} />
-          </FormItem>
-        )}
-        <FormItem label={t('common:aspectRatio')} {...formItemLayout}>
-          <RadioGroup
-            size="small"
-            defaultValue={MEDIA_ASPECT_RATIO.sixteenToNine}
-            value={aspectRatio}
-            onChange={handleAspectRatioChanged}
-            disabled={![RESOURCE_TYPE.video, RESOURCE_TYPE.none].includes(getResourceType(sourceUrl))}
-            >
-            {Object.values(MEDIA_ASPECT_RATIO).map(ratio => (
-              <RadioButton key={ratio} value={ratio}>{ratio}</RadioButton>
-            ))}
-          </RadioGroup>
-        </FormItem>
-        <FormItem label={t('common:videoDisplay')} {...formItemLayout}>
-          <Switch
-            size="small"
-            checked={showVideo}
-            onChange={handleShowVideoChanged}
-            disabled={![RESOURCE_TYPE.video, RESOURCE_TYPE.none].includes(getResourceType(sourceUrl))}
-            />
-        </FormItem>
-        <FormItem label={t('playbackRange')} {...formItemLayout}>
-          <div className="u-input-and-button">
-            <Input value={renderPlaybackRangeInfo()} readOnly />
-            <MediaRangeSelector
-              sourceUrl={getFullSourceUrl(sourceUrl)}
-              range={playbackRange}
-              onRangeChange={handlePlaybackRangeChange}
-              />
-          </div>
-        </FormItem>
+        <MainTrackEditor
+          content={content}
+          onContentChanged={handleMainTrackContentChange}
+          />
         <FormItem label={t('common:width')} {...formItemLayout}>
           <ObjectWidthSlider value={width} onChange={handleWidthChanged} />
-        </FormItem>
-        <FormItem label={t('common:copyrightNotice')} {...formItemLayout}>
-          <MarkdownInput value={copyrightNotice} onChange={handleCopyrightNoticeChanged} />
         </FormItem>
 
         <Divider className="InteractiveMediaEditor-chapterEditorDivider" plain>{t('editChapter')}</Divider>
